pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unifi-dyndns"
version = "0.1.0"
description = "Updated DynDNS records based on the IP address of a UniFi gateway."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "take" }
]
dependencies = [
  "requests>=2.32",
  "dnspython>=2.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
unifi-dyndns = "unifi_dyndns.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
